+++ src/Services/CRM/CRMServiceBuilder.php
@@ -121,6 +121,30 @@ class CRMServiceBuilder extends AbstractServiceBuilder
         return $this->serviceCache[__METHOD__];
     }
     
+    public function itemDetailsConfiguration(): Item\Service\ItemDetailsConfiguration
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new Item\Service\ItemDetailsConfiguration(
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
+    public function dealDetailsConfiguration(): Deal\Service\DealDetailsConfiguration
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new Deal\Service\DealDetailsConfiguration(
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
     public function leadDetailsConfiguration(): Lead\Service\LeadDetailsConfiguration
     {
         if (!isset($this->serviceCache[__METHOD__])) {
@@ -385,6 +409,18 @@ class CRMServiceBuilder extends AbstractServiceBuilder
         return $this->serviceCache[__METHOD__];
     }
     
+    public function leadContact(): Lead\Service\LeadContact
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new Lead\Service\LeadContact(
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
     /**
      * @return Lead\Service\LeadProductRows
      */
@@ -497,6 +533,19 @@ class CRMServiceBuilder extends AbstractServiceBuilder
                 $this->core,
                 $this->batch,
                 $this->bulkItemsReader,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+    
+    public function address(): Address\Service\Address
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new Address\Service\Address(
+                new Address\Service\Batch($this->batch, $this->log),
+                $this->core,
                 $this->log
             );
         }