+++ tests/Integration/Services/CRM/Activity/Service/ActivityTest.php
@@ -46,8 +46,11 @@ class ActivityTest extends TestCase
     use CustomBitrix24Assertions;
 
     private Activity $activityService;
+
     private Contact $contactService;
+
     private array $contactId;
+
     private array $activityId;
 
     public function testAllSystemFieldsAnnotated(): void
@@ -162,7 +165,6 @@ class ActivityTest extends TestCase
     }
 
     /**
-     * @covers Contact::fields
      * @throws BaseException
      * @throws TransportException
      */
@@ -202,7 +204,7 @@ class ActivityTest extends TestCase
             $this->activityId[] = $this->activityService->add($newActivity[$i])->getId();;
         }
 
-        $res = $this->activityService->list(
+        $activitiesResult = $this->activityService->list(
             ['ID' => 'DESC'],
             [
                 'OWNER_ID' => $contactId,
@@ -212,7 +214,7 @@ class ActivityTest extends TestCase
         );
 
 
-      $this->assertEquals(count($newActivity), count($res->getActivities()));
+      $this->assertEquals(count($newActivity), count($activitiesResult->getActivities()));
     }
 
     /**
@@ -293,15 +295,11 @@ class ActivityTest extends TestCase
         );
     }
 
-    public function tearDown(): void
+    protected function tearDown(): void
     {
-        foreach ($this->activityService->batch->delete($this->activityId) as $result) {
-        }
-        foreach ($this->contactService->batch->delete($this->contactId) as $result) {
-        }
     }
 
-    public function setUp(): void
+    protected function setUp(): void
     {
         $this->activityService = Fabric::getServiceBuilder()->getCRMScope()->activity();
         $this->contactService = Fabric::getServiceBuilder()->getCRMScope()->contact();